backend/pyproject.toml
[project]
name = "auto-structure-analysis-backend"
version = "0.1.0"
description = "FastAPI backend for automated structural analysis using PyNite and ArUco markers"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn>=0.30.0",
    "PyNiteFEA>=0.0.89",
    "opencv-python-headless>=4.10.0",
    "numpy>=2.0.0",
    "pillow>=10.4.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.8.0",
    "ultralytics>=8.0.0",
    "matplotlib>=3.7.0",
    "reportlab>=4.0.0",
    "slowapi>=0.1.9",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
